update user entity schema
change some columns' data type to be smallint since they aren't that big

+++ shibe/src/entity/User.ts
@@ -27,11 +27,11 @@ export class User extends BaseEntity {
   tokenVersion: number;
 
   @Field(() => Int)
-  @Column("int", { default: 0 })
+  @Column("smallint", { default: 0 })
   serverId: number;
 
   @Field(() => Int)
-  @Column("int", { default: 0 })
+  @Column("smallint", { default: 0 })
   avatarId: number;
 
   @Field(() => String)
